misc/mzutil/requirements.txt
psycopg2-binary==2.8.6
toml==0.10.2
